# Redesign Sidekiq internals