--- a/Gemfile
+++ b/Gemfile
@@ -4,6 +4,7 @@ gemspec
 gem 'rails', '~> 4.2'
 gem 'simplecov'
 gem 'minitest'
+gem 'toxiproxy'
 
 platforms :rbx do
   gem 'rubysl', '~> 2.0'         # if using anything in the ruby standard library