--- a/Gemfile
+++ b/Gemfile
@@ -4,12 +4,12 @@ gemspec
 gem 'rails', '~> 4.2'
 gem 'simplecov'
 gem 'minitest'
+gem 'minitest-utils'
 gem 'toxiproxy'
 
 platforms :rbx do
   gem 'rubysl', '~> 2.0'         # if using anything in the ruby standard library
   gem 'psych'                    # if using yaml
-  gem 'minitest'                 # if using minitest
   gem 'rubinius-developer_tools' # if using any of coverage, debugger, profiler
 end