… 4.0-Upgrade.md
… Changes.md
… Ent-Changes.md
… Pro-3.0-Upgrade.md
… lib/sidekiq/actor.rb
… lib/sidekiq/manager.rb
… lib/sidekiq/processor.rb
… lib/sidekiq/redis_connection.rb
… lib/sidekiq/scheduled.rb
… lib/sidekiq/util.rb
… lib/sidekiq/version.rb
… myapp/config/initializers/sidekiq.rb
… sidekiq.gemspec
… test/helper.rb
… test/test_actors.rb
… test/test_api.rb
… test/test_fetch.rb
… test/test_launcher.rb
… test/test_manager.rb
… test/test_middleware.rb
… test/test_processor.rb
… test/test_scheduled.rb