My recent experiment to remove Celluloid has reaped huge rewards.  This branch does several things:

1. Rewrite Sidekiq actors to use raw Threads and data structures.
2. Make Processor#stats asynchronous, which removes two Redis round trips per job.
3. Redesign job fetch so each Processor fetches its own job in parallel, this (along with (2)) makes the system much more resilient to higher Redis latency

Processing 100,000 no-op jobs with one process and 25 threads on MRI 2.2:

* Master: 125 sec
* Without Celluloid: 57 sec
* Without Celluloid and Async stats: 20 sec
* With 1ms of latency, sequential fetch: 7 min
* With 1ms of latency, parallel fetch: 20 sec

This work would be released as Sidekiq 4.0, with an ETA of Thanksgiving.