@mperham 2015-10-07
I reimplemented fetching so that each Processor fetches its own jobs, instead of delegating fetch to a dedicated thread.  When Redis has 1ms of latency, this reduces the time needed to process 100,000 jobs from 7min to 20sec.