@uhoh-itsmaciek 2015-10-08
Looks great. Did you by chance measure memory usage as well?