@mperham 2015-10-08
8x less garbage generated!

> On Oct 7, 2015, at 22:15, Maciek Sakrejda <notifications@github.com> wrote:
> 
> Looks great. Did you by chance measure memory usage as well?
> 
> —
> Reply to this email directly or view it on GitHub.
>