@tarcieri 2015-10-09
Have you considered using a `ThreadPoolExecutor`? This will give you a very fast and robust implementation on JRuby (now):

https://ruby-concurrency.github.io/concurrent-ruby/Concurrent/ThreadPoolExecutor.html