@tarcieri 2015-10-09
@mperham on JRuby at least, it's backed by a `java.util.concurrent.ThreadPoolExecutor` and should have both lower thread coordination and memory overhead than anything that can be done in pure Ruby