@mhenrixon 2015-10-12
Solid! Looks like you were able to delete more code than you added :+1:  Anything I need to worry about with the middleware? Couldn't see anything while going over the code.