@mperham 2015-10-12
Most public APIs should be unchanged, including middleware.  As of today. :-)

> On Oct 12, 2015, at 10:16, Mikael Henriksson <notifications@github.com> wrote:
> 
> Solid! Looks like you were able to delete more code than you added Anything I need to worry about with the middleware? Couldn't see anything while going over the code.
> 
> —
> Reply to this email directly or view it on GitHub.
>