@drewblas 2015-10-14
@mperham What happens if the status updator/heartbeat thread dies?  Will it have status info about jobs that finished but it never sends that info back to Redis?  It seems like the buffering of stats creates a hole for data to get lost in a crash, resulting in even more double-run or stuck jobs.