@mperham 2015-10-14
@drewblas I'm not sure what you are referring to.  If you can be more explicit about a particular failure case, I can explain what should happen.