@Magicdream 2015-10-15
@mperham 
> 8x less garbage generated!
Is it because of removed Celluloid?