@mperham 2015-10-22
I did a quick Resque 1.25.2 benchmark for comparison.  Time to process 10,000 noop jobs:

1 process, 96 sec
25 processes, 31 sec

Extrapolating to 100,000 jobs would be ~300 seconds.  Sidekiq does the same in 20 seconds so Sidekiq's job overhead appears to be 15x less.