@narapon 2015-10-28
Excited to try this, good work!
On 29 Oct 2015 03:05, "Tim Carey-Smith" <notifications@github.com> wrote:

> [image: :+1:]
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/mperham/sidekiq/pull/2593#issuecomment-151958303>.
>