@mperham 2015-10-29
@esbanarango You can start using it like any other beta software. https://www.reddit.com/r/sidekiq/comments/3qpxfn/sidekiq_400pre1_and_sidekiq_pro_300pre1_now/